fix auto_weaken
Previously the weakenings added were on a vertex on the control
flow graph.  Now they are on the edges.  That is to say, the
action of auto_weaken is to split the last guard edge of a
sequence of such edges into a guard part and a weaken part.

This modification makes all weaken annotations in the examples
I have superfluous.